requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.41
asyncpg==0.31.0
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
faker==40.13.0
